+++ .yarn/sdks/eslint/package.json
@@ -1,14 +1,6 @@
 {
   "name": "eslint",
-  "version": "8.51.0-sdk",
+  "version": "8.45.0-sdk",
   "main": "./lib/api.js",
-  "type": "commonjs",
-  "bin": {
-    "eslint": "./bin/eslint.js"
-  },
-  "exports": {
-    "./package.json": "./package.json",
-    ".": "./lib/api.js",
-    "./use-at-your-own-risk": "./lib/unsupported-api.js"
-  }
+  "type": "commonjs"
 }
