+++ .yarn/sdks/prettier/package.json
@@ -1,7 +1,6 @@
 {
   "name": "prettier",
-  "version": "3.0.3-sdk",
-  "main": "./index.cjs",
-  "type": "commonjs",
-  "bin": "./bin/prettier.cjs"
+  "version": "3.0.0-sdk",
+  "main": "./index.js",
+  "type": "commonjs"
 }
