+++ .yarn/sdks/typescript/package.json
@@ -1,10 +1,6 @@
 {
   "name": "typescript",
-  "version": "5.2.2-sdk",
+  "version": "5.1.6-sdk",
   "main": "./lib/typescript.js",
-  "type": "commonjs",
-  "bin": {
-    "tsc": "./bin/tsc",
-    "tsserver": "./bin/tsserver"
-  }
+  "type": "commonjs"
 }
