+++ packages/apollo-shared/src/Changes/index.ts
@@ -11,7 +11,6 @@ import { DiscontinuousLocationStartChange } from './DiscontinuousLocationStartCh
 import { FeatureAttributeChange } from './FeatureAttributeChange'
 import { LocationEndChange } from './LocationEndChange'
 import { LocationStartChange } from './LocationStartChange'
-import { StrandChange } from './StrandChange'
 import { TypeChange } from './TypeChange'
 import { UserChange } from './UserChange'
 
@@ -29,7 +28,6 @@ export const changes = {
   FeatureAttributeChange,
   LocationEndChange,
   LocationStartChange,
-  StrandChange,
   TypeChange,
   UserChange,
 }
@@ -47,6 +45,5 @@ export * from './DiscontinuousLocationStartChange'
 export * from './FeatureAttributeChange'
 export * from './LocationEndChange'
 export * from './LocationStartChange'
-export * from './StrandChange'
 export * from './TypeChange'
 export * from './UserChange'
