+++ packages/apollo-shared/src/util.ts
@@ -94,15 +94,3 @@ export function makeGFF3Feature(
       : [],
   }))
 }
-
-export function splitStringIntoChunks(
-  input: string,
-  chunkSize: number,
-): string[] {
-  const chunks: string[] = []
-  for (let i = 0; i < input.length; i += chunkSize) {
-    const chunk = input.slice(i, i + chunkSize)
-    chunks.push(chunk)
-  }
-  return chunks
-}
