+++ packages/jbrowse-plugin-apollo/package.json
@@ -76,7 +76,7 @@
   },
   "peerDependencies": {
     "@jbrowse/core": "^2.7.0",
-    "@mui/material": "^5.11.14",
+    "@mui/material": "^5.11.10",
     "mobx": "^6.6.1",
     "mobx-react": "^7.2.1",
     "mobx-state-tree": "^5.1.7",
@@ -91,8 +91,8 @@
     "@jbrowse/core": "^2.7.0",
     "@jbrowse/development-tools": "^2.1.1",
     "@jest/globals": "^29.0.3",
-    "@mui/material": "^5.11.14",
-    "@mui/x-data-grid": "^7.0.0",
+    "@mui/material": "^5.11.10",
+    "@mui/x-data-grid": "^6.0.1",
     "@types/autosuggest-highlight": "^3",
     "@types/file-saver": "^2",
     "@types/node": "^18.14.2",
