--- packages/jbrowse-plugin-apollo/src/FeatureDetailsWidget/index.ts
@@ -1,2 +1,0 @@
-export { ApolloFeatureDetailsWidget } from './ApolloFeatureDetailsWidget'
-export { ApolloFeatureDetailsWidgetModel } from './model'
