+++ packages/jbrowse-plugin-apollo/src/components/ViewChangeLog.tsx
@@ -76,11 +76,9 @@ export function ViewChangeLog({ handleClose, session }: ViewChangeLogProps) {
       headerName: 'Change JSON',
       width: 600,
       renderCell: ({ value }) => (
-        <textarea
-          className={classes.changeTextarea}
-          value={JSON.stringify(value)}
-          readOnly
-        />
+        <textarea className={classes.changeTextarea} readOnly>
+          {JSON.stringify(value)}
+        </textarea>
       ),
       valueFormatter: ({ value }) => JSON.stringify(value),
     },
@@ -90,7 +88,7 @@ export function ViewChangeLog({ handleClose, session }: ViewChangeLogProps) {
       headerName: 'Time',
       width: 160,
       type: 'dateTime',
-      valueGetter: (value) => value && new Date(value),
+      valueGetter: ({ value }) => value && new Date(value),
     },
   ]
