+++ packages/jbrowse-plugin-apollo/src/makeDisplayComponent.tsx
@@ -2,7 +2,8 @@ import PluginManager from '@jbrowse/core/PluginManager'
 import type LinearGenomeViewPlugin from '@jbrowse/plugin-linear-genome-view'
 import ExpandLessIcon from '@mui/icons-material/ExpandLess'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import { Typography, alpha } from '@mui/material'
+import { Typography } from '@mui/material'
+import { alpha } from '@mui/material/styles'
 import { observer } from 'mobx-react'
 import React, { useCallback, useEffect, useRef } from 'react'
 import { makeStyles } from 'tss-react/mui'
